overlapping area - functional
